+++ lib/logstash/plugin_mixins/ecs_compatibility_support.rb
@@ -25,17 +25,36 @@ module LogStash
 
         # If our base does not include an `ecs_compatibility` config option,
         # include the legacy adapter to ensure it gets defined.      
-        base.send(:include, LegacyAdapter) unless base.get_config.include?("ecs_compatibility")
+        base.send(:include, LegacyAdapter) unless base.method_defined?(:ecs_compatibility?)
       end
 
       ##
-      # Declares a boolean `ecs_compatibility` config option on the `base` that
-      # defaults to `false`.
-      #
+      # This `ECSCompatibilitySupport` cannot be extended into an existing object.
       # @api private
+      #
+      # @param base [Object]
+      # @raise [ArgumentError]
+      def self.extended(base)
+        fail(ArgumentError, "`#{self}` cannot be extended into an existing object.")
+      end
+
+      ##
+      # Implements `ecs_compatibility?` method backed by a boolean `ecs_compatibility`
+      # config option that defaults to `false`.
+      #
+      # @api internal
       module LegacyAdapter
         def self.included(base)
           base.config(:ecs_compatibility, :validate => :boolean, :default => false)
+        end
+
+        ##
+        # @api public
+        # @return [Boolean]
+        def ecs_compatibility?
+          # NOTE: The @ecs_compatibility instance variable is an implementation detail of
+          #       this `LegacyAdapter` and plugins MUST NOT rely in its presence or value.
+          @ecs_compatibility
         end
       end
     end