+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.0.0
 
- - Implements boolean `ecs_compatibility` config option, unless already provided by Logstash core.
+ - Support Mixin for ensuring a plugin has an `ecs_compatibility?` method that is configurable from a boolean `ecs_compatibility` option, using the implementation from Logstash core if available.